requirements.txt
# LangChain 核心库（1.x 生态拆分包）
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# 向量存储
faiss-cpu>=1.7.4

# Embedding（使用 OpenAI 兼容 API）
langchain-openai>=0.0.5

# 本地 Embedding（如果 API 不支持 embedding，使用此选项）
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0

# 环境变量管理
python-dotenv>=1.0.0

# 文档加载（如果需要支持更多格式，可以添加）
# pypdf>=3.0.0  # PDF 支持
# python-docx>=1.0.0  # Word 支持

# Web API
fastapi>=0.111.0
uvicorn>=0.30.0

